moved util.py into core
