+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 - 0.2.9
   - (TODO) change sign in gradient simulation - flips orientation of old reconstructions
-  - (TODO) Included util file in mr0
+  - Included util file in mr0
   - (TODO) Improved plot function to support pypulseq 1.4.0-dev
 - 0.2.8
   - Updated playground notebook to use new FOV definitions
